Dev enzo

+++ backend/src/controllers/answers.controller.js
@@ -2,6 +2,8 @@ import prisma from "../prismaClient.js";
 import { uploadToCloudinary } from "../../services/cloudinary.service.js";
 import { error } from "node:console";
 import { CreateAnswersSchema, UpdateAnswersSchema } from "../validators/answers.schema.js";
+import { sendAnswerReceivedEmail} from "../services/email.service.js";
+import { hmacEmail, encryptEmail, decryptEmail } from "../utils/emailCrypto.js";
 
 export const createAnswer = async (req, res) => {
   try {
@@ -14,8 +16,10 @@ export const createAnswer = async (req, res) => {
     }
 
     const data = result.data;
+
     const depo = req.depo;
     if (!depo) return res.status(404).json({ error: "Depo not found" });
+
     const answer = await prisma.T_Answers.create({
       data: {
         Text_Answer: data.description,
@@ -23,6 +27,29 @@ export const createAnswer = async (req, res) => {
         ID_User: req.user.userId,
       },
     });
+
+    const owner = await prisma.T_Users.findUnique({
+      where: {
+        ID_User: depo.ID_User,
+      },
+      select: {
+        Email_Encrypted_User: true,
+        Login_User: true,
+      },
+    });
+
+    const sender = await prisma.t_Users.findUnique({
+      where: {
+        ID_User: answer.ID_User,
+      },
+      select: {
+        Login_User: true,
+      }
+    })
+
+    const email = decryptEmail(owner.Email_Encrypted_User);
+
+    await sendAnswerReceivedEmail(email, sender, depo, answer);
     res.status(201).json(answer);
   } catch (err) {
     console.error(err);
@@ -32,6 +59,15 @@ export const createAnswer = async (req, res) => {
 
 export const updateAnswer = async (req, res) => {
   try {
+    const answer = req.answer;
+    if (!answer) return res.status(404).json({ error: "answer not found" });
+
+    if (req.user.userId !== answer.ID_User && req.user.role !== "ADMIN") {
+      return res.status(403).json({
+        error: "Forbidden",
+      });
+    }
+
     const result = UpdateAnswersSchema.safeParse(req.body);
 
     if (!result.success) {
@@ -62,6 +98,12 @@ export const deleteAnswer = async (req, res) => {
   try {
     const answer = req.answer;
     if (!answer) return res.status(404).json({ error: "Answer not found" });
+
+    if (req.user.userId !== answer.ID_User && req.user.role !== "ADMIN") {
+      return res.status(403).json({
+        error: "Forbidden",
+      });
+    }
 
     await prisma.T_Answers.delete({ where: { ID_Answer: req.params.id } });
     res.json({ message: "Deleted" });
